backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
requests-oauthlib==1.3.1
requests==2.31.0
deepface==0.0.93
numpy==1.26.4
scikit-learn==1.4.2
httpx==0.27.0
python-multipart==0.0.9
itsdangerous==2.2.0
apscheduler==3.10.4
Pillow==10.3.0
